create remote-modal.js
