+++ src/js/test-render-all-books.js
@@ -1,6 +1,5 @@
-import { receiveBookByCategory } from './category.js';
 import { renderTopBooks } from './render-top-books.js';
-import { renderModal } from './modal-book-selection';
+import { openModal } from './remote-modal';
 
 import BookApi from './services.js';
 
@@ -17,9 +16,6 @@ bookApi
 
 const allBooksListEl = document.querySelector('.all-book-list');
 
-const modalBackdropEl = document.querySelector('.backdrop');
-
-// ============================================================================
 allBooksListEl.addEventListener('click', e => {
   e.preventDefault();
 
@@ -36,48 +32,3 @@ allBooksListEl.addEventListener('click', e => {
     openModal(bookItem.dataset.bookId);
   }
 });
-// ============================================================================
-
-function openModal(bookId) {
-  bookApi.getBookDetail(bookId).then(bookDetails => {
-    renderModal(bookDetails);
-
-    modalBackdropEl.classList.remove('is-hidden');
-
-    const closeButton = modalBackdropEl.querySelector('.close');
-    const modal = modalBackdropEl.querySelector('#modal');
-
-    const closeModal = () => {
-      modalBackdropEl.classList.add('is-hidden');
-      removeEventListener('click', event1);
-      removeEventListener('click', event2);
-      removeEventListener('keydown', event3);
-    };
-
-    const event1 = closeButton.addEventListener('click', () => {
-      closeModal();
-    });
-
-    const event2 = document.addEventListener('click', e => {
-      if (!modal.contains(e.target)) {
-        closeModal();
-      }
-    });
-
-    const event3 = document.addEventListener('keydown', e => {
-      if (e.key === 'Escape') {
-        closeModal();
-      }
-    });
-  });
-}
-
-// bookApi
-//   .getBooksByCategory('Audio Fiction')
-//   .then(category => {
-//     console.log(category);
-//     receiveBookByCategory(category);
-//   })
-//   .catch(err => {
-//     console.log(err);
-//   });
